refactor: follow the current sample style
